refactor: checkstyle 만족하도록 변경
- EmptyLineSeparator 의 현재 설정은
--- import 사이의 1줄의 emptyline 은 허용하지만, 2줄 이상은 허용하지 않음
--- token 을 통해 해당 설정을 적용할 대상들을 명시해 둠
--- package 와 import 사이의 하나의 emptyline 도 허용하고 있음
