+++ ynoolee/src/test/java/ch05/ch5_5/v3/IdGeneratorV3ImplTest.java
@@ -41,7 +41,6 @@ class IdGeneratorV3ImplTest {
         }
     }
 
-
     @ParameterizedTest
     @DisplayName("주어진 길이의 랜덤문자열을 생성한다")
     @ValueSource(ints = {3, 4, 5})
